include the WSClient contract
