+++ packages/core/src/testUtils.ts
@@ -33,7 +33,6 @@ export const configureJestStore = (
       project: PROJECT_ID,
       token: TOKEN,
       devTools: false,
-      emitter: new EventEmitter(),
     },
     SessionConstructor: BaseSession,
     runRootSaga: false,
@@ -61,7 +60,6 @@ export const configureFullStack = () => {
       project: PROJECT_ID,
       token: TOKEN,
       devTools: false,
-      emitter,
     },
     SessionConstructor: jest.fn().mockImplementation(() => {
       return session
