+++ packages/js/src/fabric/Conversation.ts
@@ -24,7 +24,7 @@ import type {
   JoinConversationResponse,
   JoinConversationResult,
   CoversationSubscribeResult,
-} from './types'
+} from './interfaces'
 import { conversationWorker } from './workers'
 import { buildPaginatedResult } from '../utils/paginatedResult'
 import { makeQueryParamsUrls } from '../utils/makeQueryParamsUrl'
