+++ packages/js/src/fabric/createWSClient.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@signalwire/core'
 import { SATSession } from './SATSession'
-import { WSClientOptions } from './types'
+import { WSClientOptions } from './interfaces'
 
 export const createWSClient = (options: WSClientOptions) => {
   const store = configureStore({
