+++ packages/js/src/fabric/interfaces/device.ts
@@ -1,0 +1,21 @@
+export type RegisterDeviceType = 'iOS' | 'Android' | 'Desktop'
+
+export interface RegisterDeviceParams {
+  deviceType: RegisterDeviceType
+  deviceToken: string
+}
+
+export interface UnregisterDeviceParams {
+  id: string
+}
+
+export interface RegisterDeviceResponse {
+  date_registered: Date
+  device_name?: string
+  device_token: string
+  device_type: RegisterDeviceType
+  id: string
+  push_notification_key: string
+}
+
+export type RegisterDeviceResult = RegisterDeviceResponse
