--- packages/js/src/fabric/types.ts
@@ -1,402 +1,0 @@
-import type { ConversationEventParams, UserOptions } from '@signalwire/core'
-import { HTTPClient } from './HTTPClient'
-import { WSClient } from './WSClient'
-import { Conversation } from './Conversation'
-import type { FabricRoomSession } from './FabricRoomSession'
-
-export interface SignalWireClientParams extends WSClientOptions {}
-
-export interface SignalWireContract {
-  httpHost: HTTPClient['httpHost']
-  registerDevice: HTTPClient['registerDevice']
-  unregisterDevice: HTTPClient['unregisterDevice']
-  getSubscriberInfo: HTTPClient['getSubscriberInfo']
-  disconnect: WSClient['disconnect']
-  online: WSClient['online']
-  offline: WSClient['offline']
-  dial: WSClient['dial']
-  reattach: WSClient['reattach']
-  handlePushNotification: WSClient['handlePushNotification']
-  updateToken: WSClient['updateToken']
-  address: {
-    getAddresses: HTTPClient['getAddresses']
-    getAddress: HTTPClient['getAddress']
-  }
-  conversation: {
-    getConversations: Conversation['getConversations']
-    getMessages: Conversation['getMessages']
-    getConversationMessages: Conversation['getConversationMessages']
-    subscribe: Conversation['subscribe']
-    sendMessage: Conversation['sendMessage']
-    join: Conversation['joinConversation']
-  }
-  chat: {
-    getMessages: Conversation['getChatMessages']
-    subscribe: Conversation['subscribeChatMessages']
-    sendMessage: Conversation['sendMessage']
-    join: Conversation['joinConversation']
-  }
-}
-
-export type SignalWireClient = SignalWireContract
-
-/**
- * WSClient
- */
-
-export interface OnlineParams {
-  incomingCallHandlers: IncomingCallHandlers
-}
-
-export interface PushNotificationPayload {
-  encryption_type: 'aes_256_gcm'
-  notification_uuid: string
-  with_video: 'true' | 'false'
-  incoming_caller_name: string
-  incoming_caller_id: string
-  tag: string
-  invite: string
-  title: string
-  type: 'call_invite'
-  iv: string
-  version: string
-  decrypted: Record<string, any>
-}
-
-export interface CallParams {
-  /** HTML element in which to display the video stream */
-  rootElement?: HTMLElement | null
-  /** Disable ICE UDP transport policy */
-  disableUdpIceServers?: boolean
-  /** Audio constraints to use when joining the room. Default: `true`. */
-  audio?: MediaStreamConstraints['audio']
-  /** Video constraints to use when joining the room. Default: `true`. */
-  video?: MediaStreamConstraints['video']
-  /** Negotiate the incoming audio from the RTC. Default: `true`. */
-  negotiateAudio?: boolean
-  /** Negotiate the incoming video from the RTC. Default: `true`. */
-  negotiateVideo?: boolean
-  /** User & UserAgent metadata */
-  userVariables?: WSClientOptions['userVariables']
-}
-
-export interface DialParams extends CallParams {
-  to: string
-  nodeId?: string
-}
-
-export type FabricUserOptions = Omit<UserOptions, 'onRefreshToken'>
-
-export interface WSClientOptions extends FabricUserOptions {
-  /** HTML element in which to display the video stream */
-  rootElement?: HTMLElement
-  /** Call back function to receive the incoming call */
-  incomingCallHandlers?: IncomingCallHandlers
-  /** User & UserAgent metadata */
-  userVariables?: Record<string, any>
-}
-
-/**
- * Incoming Call Manager
- */
-
-export type IncomingInviteSource = 'websocket' | 'pushNotification'
-
-export interface IncomingInvite {
-  callID: string
-  sdp: string
-  caller_id_name: string
-  caller_id_number: string
-  callee_id_name: string
-  callee_id_number: string
-  display_direction: string
-  nodeId: string
-}
-
-export interface IncomingInviteWithSource extends IncomingInvite {
-  source: IncomingInviteSource
-}
-
-export interface IncomingCallNotification {
-  invite: {
-    details: IncomingInvite
-    accept: (param: CallParams) => Promise<FabricRoomSession>
-    reject: () => Promise<void>
-  }
-}
-export type IncomingCallHandler = (
-  notification: IncomingCallNotification
-) => void
-
-export interface IncomingCallHandlers {
-  all?: IncomingCallHandler
-  pushNotification?: IncomingCallHandler
-  websocket?: IncomingCallHandler
-}
-
-/**
- * Paginated response and result
- */
-
-export interface PaginatedResponse<T> {
-  data: Array<T>
-  links: {
-    first?: string
-    self?: string
-    next?: string
-    prev?: string
-  }
-}
-
-export interface PaginatedResult<T> {
-  data: Array<T>
-  self(): Promise<PaginatedResult<T> | undefined>
-  nextPage(): Promise<PaginatedResult<T> | undefined>
-  prevPage(): Promise<PaginatedResult<T> | undefined>
-  firstPage(): Promise<PaginatedResult<T> | undefined>
-  hasNext: boolean
-  hasPrev: boolean
-}
-
-/**
- * Addresses
- */
-
-export type ResourceType = 'app' | 'call' | 'room' | 'subscriber'
-
-export interface GetAddressResponse {
-  id: string
-  display_name: string
-  name: string
-  preview_url?: string
-  cover_url?: string
-  resource_id: string
-  type: ResourceType
-  channels: {
-    audio?: string
-    messaging?: string
-    video?: string
-  }
-  locked: boolean
-  created_at: string
-}
-
-export type Address = GetAddressResponse
-
-export interface GetAddressesParams {
-  type?: string
-  displayName?: string
-  pageSize?: number
-  sortBy?: 'name' | 'created_at'
-  sortOrder?: 'asc' | 'desc'
-}
-
-export interface GetAddressByIdParams {
-  id: string
-}
-
-export interface GetAddressByNameParams {
-  name: string
-}
-
-export type GetAddressParams = GetAddressByIdParams | GetAddressByNameParams
-
-export type GetAddressResult = GetAddressResponse
-
-export type GetAddressesResponse = PaginatedResponse<GetAddressResponse>
-
-export type GetAddressesResult = PaginatedResult<GetAddressResponse>
-
-/**
- * Conversations
- */
-export interface ConversationContract {
-  readonly addressId: string
-  readonly createdAt: number
-  readonly id: string
-  readonly lastMessageAt: number
-  readonly metadata: Record<string, any>
-  readonly name: string
-  sendMessage(
-    params: ConversationAPISendMessageParams
-  ): Promise<SendConversationMessageResult>
-  getMessages(
-    params?: ConversationAPIGetMessagesParams
-  ): Promise<GetConversationMessagesResult>
-}
-
-export interface SendConversationMessageParams {
-  text: string
-  addressId: string
-  metadata?: Record<string, any>
-  details?: Record<string, any>
-}
-
-export interface SendConversationMessageResponse {
-  table: {
-    conversation_id: string
-    text: string
-  }
-}
-
-export type SendConversationMessageResult = SendConversationMessageResponse
-
-export interface GetConversationsParams {
-  pageSize?: number
-}
-
-export interface ConversationResponse {
-  address_id: string
-  created_at: number
-  id: string
-  last_message_at: number
-  metadata: Record<string, any>
-  name: string
-}
-
-export type GetConversationsResponse = PaginatedResponse<ConversationResponse>
-
-export type GetConversationsResult = PaginatedResult<ConversationContract>
-
-export type CoversationSubscribeCallback = (
-  event: ConversationEventParams
-) => unknown
-
-export interface CoversationSubscribeResult {
-  unsubscribe: () => void
-}
-
-export interface ConversationChatMessagesSubscribeParams {
-  addressId: string
-  onMessage: CoversationSubscribeCallback
-}
-
-export type ConversationChatMessagesSubscribeResult = CoversationSubscribeResult
-
-export interface JoinConversationParams {
-  addressId: string
-}
-
-export interface JoinConversationResponse {
-  table: {
-    conversation_id: string
-    text: string
-  }
-}
-
-export type JoinConversationResult = JoinConversationResponse
-
-/**
- * Conversation Messages
- */
-export interface GetMessagesParams {
-  pageSize?: number
-}
-
-export interface ConversationMessage {
-  id: string
-  address_id: string
-  conversation_id: string
-  user_id: string
-  from_address_id: string
-  ts: number
-  metadata?: Record<string, any>
-  details: Record<string, any>
-  type: string
-  subtype: string
-  kind?: string
-  text?: string
-}
-
-export type GetMessagesResult = PaginatedResult<ConversationMessage>
-
-export type ConversationChatMessage = Omit<ConversationMessage, 'kind'> & {
-  text: string
-}
-
-export interface GetConversationChatMessageParams {
-  addressId: string
-  pageSize?: number
-}
-
-export type GetConversationChatMessageResult =
-  PaginatedResult<ConversationChatMessage>
-
-export interface GetConversationMessagesResponse
-  extends PaginatedResponse<ConversationMessage> {}
-
-export interface GetConversationMessagesParams {
-  addressId: string
-  pageSize?: number
-}
-
-export type GetConversationMessagesResult = PaginatedResult<ConversationMessage>
-
-/**
- * Conversation API
- */
-export interface ConversationAPISendMessageParams {
-  text: string
-}
-
-export interface ConversationAPIGetMessagesParams {
-  pageSize?: number
-}
-
-/**
- * Subscriber info
- */
-export interface GetSubscriberInfoResponse {
-  id: string
-  email: string
-  first_name?: string
-  last_name?: string
-  display_name?: string
-  job_title?: string
-  time_zone?: number
-  country?: string
-  region?: string
-  company_name?: string
-  push_notification_key: string
-  app_settings?: {
-    display_name: string
-    scopes: string[]
-  }
-  fabric_addresses: Address[]
-}
-
-export type GetSubscriberInfoResult = GetSubscriberInfoResponse
-
-/**
- * Device registration
- */
-export type RegisterDeviceType = 'iOS' | 'Android' | 'Desktop'
-
-export interface RegisterDeviceParams {
-  deviceType: RegisterDeviceType
-  deviceToken: string
-}
-
-export interface UnregisterDeviceParams {
-  id: string
-}
-
-export interface RegisterDeviceResponse {
-  date_registered: Date
-  device_name?: string
-  device_token: string
-  device_type: RegisterDeviceType
-  id: string
-  push_notification_key: string
-}
-
-export type RegisterDeviceResult = RegisterDeviceResponse
-
-export { FabricRoomSession }
-export type {
-  ConversationMessageEventName,
-  ConversationMessageEventParams,
-  ConversationMessageEvent,
-  ConversationEvent,
-  ConversationEventParams,
-} from '@signalwire/core'
