+++ packages/js/src/testUtils.ts
@@ -16,7 +16,6 @@ export const configureJestStore = () => {
       project: PROJECT_ID,
       token: TOKEN,
       devTools: false,
-      emitter: new EventEmitter(),
     },
     SessionConstructor: JWTSession,
     runRootSaga: false,
@@ -43,7 +42,6 @@ export const configureFullStack = () => {
       project: PROJECT_ID,
       token: TOKEN,
       devTools: false,
-      emitter,
     },
     SessionConstructor: jest.fn().mockImplementation(() => {
       return session
